Upgrade to Laravel 13 on PHP 8.4
Hatchery was running Laravel 8 on PHP 8.1 with a Laravel 5 era skeleton.
This moves the stack to Laravel 13, restructures the application on the
slim skeleton introduced in Laravel 11, and replaces the toolchain pieces
that have since been abandoned.

Framework and skeleton:
- laravel/framework ^8.0 -> ^13.0, php ^8.1 -> ^8.4
- bootstrap/app.php now configures routing, middleware, scheduling and
  exception handling; app/Http/Kernel.php, app/Console/Kernel.php and
  app/Exceptions/Handler.php are gone
- providers are listed in bootstrap/providers.php; the Auth, Event,
  Broadcast and Route service providers are replaced by auto-discovery
  and withRouting()
- routes use [Controller::class, 'method'] instead of the controller
  namespace prefix removed in Laravel 9
- config/ keeps only what differs from the framework defaults; env keys
  follow the current names (CACHE_STORE, QUEUE_CONNECTION,
  BROADCAST_CONNECTION, MAIL_MAILER)
- the swagger info block and applicationVersion() move off the HTTP
  kernel to the base controller and App\Support\Version

Dependencies:
- laravelcollective/html (abandoned) -> rdx/laravelcollective-html, so
  the Form:: calls in the blade templates are untouched
- facade/ignition -> spatie/laravel-ignition
- fruitcake/laravel-cors dropped; Laravel handles CORS itself
- fzaninotto/faker -> fakerphp/faker, nunomaduro/larastan ->
  larastan/larastan
- livewire removed: the app had no components, only the asset tags
- codeception, codacy/coverage, jorijn/laravel-security-checker,
  doctrine/dbal, vimeo/psalm and composer/composer dropped
- intervention/image v2 -> v4, used through Laravel 13's Image facade
- pest 1 -> 5, phpunit 9 -> 13

Frontend:
- laravel-mix/webpack -> Vite, with @vite() in the layouts
- the CodeMirror AMD require() calls become static ESM imports
- .eslintrc.js -> eslint.config.js flat config, yarn.lock ->
  package-lock.json

Fixes found along the way:
- Blade 13 registers a @context directive, which broke the JSON-LD blocks
  in the layouts and on the project page; the keys are now escaped
- File::isValidIcon() kept working for base64 stored content, which
  Intervention v2 detected automatically and the new API does not
- Project::$dates was removed in Laravel 10, so published_at moves to
  $casts
- a migration called DB::getDoctrineSchemaManager(), gone since Laravel 11
- implicit nullable parameters, deprecated on PHP 8.4

The OpenAPI annotations are now PHP attributes: l5-swagger 11 only reads
attributes by default, and the alternative was keeping the abandoned
doctrine/annotations package plus an analyser object in config, which
would have broken config:cache. The generated api-docs.json is byte for
byte identical to the one the docblocks produced.

phpstan stays at level 8. Larastan 3 is much stricter than the 0.6 the
project pinned, so relation generics, factory generics and list<string>
property types are spelled out across the models and factories.

CI runs on PHP 8.4/8.5 and Node 22/24 with npm and composer audit, and
gains workflows for phpcs, markdownlint and a Dockerfile build. Travis
and Scrutinizer config is removed; neither has run here for years.

245 tests pass, phpstan and phpcs are clean.

+++ app/Support/Helpers.php
@@ -74,7 +74,7 @@ class Helpers
     {
         $units = ['B', 'KiB', 'MiB', 'GiB', 'TiB'];
         $bytes = max($bytes, 0);
-        $pow = floor(($bytes ? log($bytes) : 0) / log(1024));
+        $pow = (int) floor(($bytes ? log($bytes) : 0) / log(1024));
         $pow = min($pow, count($units) - 1);
         $bytes /= (1 << (10 * $pow));
